# internal/web/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: qui API
  description: qui API for managing multiple qBittorrent instances
  version: 1.0.0
  license:
    name: GPL-2.0-or-later
    url: https://opensource.org/licenses/GPL-2.0
servers:
  - url: http://localhost:7476
    description: Local development server
  - url: /
    description: Current server
security:
  - ApiKeyAuth: []
  - SessionAuth: []
paths:
...
  /api/dir-scan/directories/{directoryID}/scan:
    post:
      tags:
        - Dir Scan
      summary: Trigger a directory scan
      description: Starts a manual scan run for the given directory
      parameters:
        - $ref: '#/components/parameters/directoryID'
      responses:
        '202':
          description: Scan started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirScanTriggerResponse'
        '404':
          description: Directory not found
        '409':
          description: A scan is already in progress
        '500':
          description: Failed to start scan
    delete:
      tags:
        - Dir Scan
      summary: Cancel a directory scan
      description: Cancels the currently running scan for the given directory, if any
      parameters:
        - $ref: '#/components/parameters/directoryID'
      responses:
        '204':
          description: Scan canceled (or no scan running)
        '404':
          description: Directory not found
        '500':
          description: Failed to cancel scan
...
  /api/instances/{instanceID}/orphan-scan/scan:
    post:
      tags:
        - Orphan Scan
      summary: Trigger orphan scan
      description: Start a manual orphan file scan for an instance. Returns immediately with the run ID.
      parameters:
        - $ref: '#/components/parameters/instanceID'
      responses:
        '202':
          description: Scan started
          content:
            application/json:
              schema:
                type: object
                properties:
                  runId:
                    type: integer
                    format: int64
                    description: ID of the created scan run
        '403':
          description: Instance does not have local filesystem access enabled
        '404':
          description: Instance not found
        '409':
          description: A scan is already in progress for this instance
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    SessionAuth:
      type: apiKey
      in: cookie
      name: user_session
      description: Session cookie authentication
  parameters:
    instanceID:
      name: instanceID
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: qBittorrent instance ID
...
    directoryID:
      name: directoryID
      in: path
      required: true
      schema:
        type: integer
      description: Directory scan directory ID
...
  schemas:
...
    DirScanTriggerResponse:
      type: object
      properties:
        runId:
          type: integer
      required:
        - runId
...